refactor : change directory and rename files
