+++ src/main/java/app/bottlenote/user/repository/OauthRepository.java
@@ -1,10 +1,10 @@
-package app.bottlenote.oauth.repository;
+package app.bottlenote.user.repository;
 
 import app.bottlenote.user.domain.Users;
-import app.bottlenote.oauth.constant.SocialType;
+import app.bottlenote.user.constant.SocialType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface OauthLoginRepository extends JpaRepository<Users, Long> {
+public interface OauthRepository extends JpaRepository<Users, Long> {
 	Optional<Users> findByEmailAndSocialType(String email, SocialType socialType);
 }
